full module restructure
- full module restructure for easy of use
- some modules were removed to module repository
- simplified module include
- new module declaration and initialization macro
