+++ uCNC/cnc_hal_config.h
@@ -229,7 +229,7 @@ extern "C"
 #define TMC_SPI 2
 
 // uncomment to enable trinamic driver
-//#define STEPPER0_HAS_TMC
+// #define STEPPER0_HAS_TMC
 #ifdef STEPPER0_HAS_TMC
 #define STEPPER0_DRIVER_TYPE 2208
 // choose the interface type
