+++ uCNC/src/cnc_build.h
@@ -25,7 +25,7 @@ extern "C"
 #endif
 
 #define CNC_MAJOR_MINOR_VERSION "1.4"
-#define CNC_PATCH_VERSION ".5"
+#define CNC_PATCH_VERSION ".6pre"
 
 #define CNC_VERSION CNC_MAJOR_MINOR_VERSION CNC_PATCH_VERSION
