+++ uCNC/src/modules/digimstep.c
@@ -1,5 +1,5 @@
 /*
-    Name: digipins.c
+    Name: digimstep.c
     Description: Digital pin stepper settings module for µCNC.
 
     Copyright: Copyright (c) João Martins
@@ -19,57 +19,6 @@
 #include "../cnc.h"
 
 #ifdef ENABLE_DIGITAL_MSTEP
-void digimstep_init(void)
-{
-#ifdef STEPPER0_MSTEP0
-    io_set_output(STEPPER0_MSTEP0, (STEPPER0_MSTEP & 0x01));
-#endif
-#ifdef STEPPER0_MSTEP1
-    io_set_output(STEPPER0_MSTEP1, (STEPPER0_MSTEP & 0x02));
-#endif
-#ifdef STEPPER1_MSTEP0
-    io_set_output(STEPPER1_MSTEP0, (STEPPER1_MSTEP & 0x01));
-#endif
-#ifdef STEPPER1_MSTEP1
-    io_set_output(STEPPER1_MSTEP1, (STEPPER1_MSTEP & 0x02));
-#endif
-#ifdef STEPPER2_MSTEP0
-    io_set_output(STEPPER2_MSTEP0, (STEPPER2_MSTEP & 0x01));
-#endif
-#ifdef STEPPER2_MSTEP1
-    io_set_output(STEPPER2_MSTEP1, (STEPPER2_MSTEP & 0x02));
-#endif
-#ifdef STEPPER3_MSTEP0
-    io_set_output(STEPPER3_MSTEP0, (STEPPER3_MSTEP & 0x01));
-#endif
-#ifdef STEPPER3_MSTEP1
-    io_set_output(STEPPER3_MSTEP1, (STEPPER3_MSTEP & 0x02));
-#endif
-#ifdef STEPPER4_MSTEP0
-    io_set_output(STEPPER4_MSTEP0, (STEPPER4_MSTEP & 0x01));
-#endif
-#ifdef STEPPER4_MSTEP1
-    io_set_output(STEPPER4_MSTEP1, (STEPPER4_MSTEP & 0x02));
-#endif
-#ifdef STEPPER5_MSTEP0
-    io_set_output(STEPPER5_MSTEP0, (STEPPER5_MSTEP & 0x01));
-#endif
-#ifdef STEPPER5_MSTEP1
-    io_set_output(STEPPER5_MSTEP1, (STEPPER5_MSTEP & 0x02));
-#endif
-#ifdef STEPPER6_MSTEP0
-    io_set_output(STEPPER6_MSTEP0, (STEPPER6_MSTEP & 0x01));
-#endif
-#ifdef STEPPER6_MSTEP1
-    io_set_output(STEPPER6_MSTEP1, (STEPPER6_MSTEP & 0x02));
-#endif
-#ifdef STEPPER7_MSTEP0
-    io_set_output(STEPPER7_MSTEP0, (STEPPER7_MSTEP & 0x01));
-#endif
-#ifdef STEPPER7_MSTEP1
-    io_set_output(STEPPER7_MSTEP1, (STEPPER7_MSTEP & 0x02));
-#endif
-}
 
 /*custom gcode commands*/
 #if defined(ENABLE_PARSER_MODULES)
@@ -283,4 +232,61 @@ uint8_t m351_exec(parser_state_t *new_state, parser_words_t *words, parser_cmd_e
 }
 
 #endif
+
+DECL_MODULE(digimstep)
+{
+#ifdef STEPPER0_MSTEP0
+    io_set_output(STEPPER0_MSTEP0, (STEPPER0_MSTEP & 0x01));
+#endif
+#ifdef STEPPER0_MSTEP1
+    io_set_output(STEPPER0_MSTEP1, (STEPPER0_MSTEP & 0x02));
+#endif
+#ifdef STEPPER1_MSTEP0
+    io_set_output(STEPPER1_MSTEP0, (STEPPER1_MSTEP & 0x01));
+#endif
+#ifdef STEPPER1_MSTEP1
+    io_set_output(STEPPER1_MSTEP1, (STEPPER1_MSTEP & 0x02));
+#endif
+#ifdef STEPPER2_MSTEP0
+    io_set_output(STEPPER2_MSTEP0, (STEPPER2_MSTEP & 0x01));
+#endif
+#ifdef STEPPER2_MSTEP1
+    io_set_output(STEPPER2_MSTEP1, (STEPPER2_MSTEP & 0x02));
+#endif
+#ifdef STEPPER3_MSTEP0
+    io_set_output(STEPPER3_MSTEP0, (STEPPER3_MSTEP & 0x01));
+#endif
+#ifdef STEPPER3_MSTEP1
+    io_set_output(STEPPER3_MSTEP1, (STEPPER3_MSTEP & 0x02));
+#endif
+#ifdef STEPPER4_MSTEP0
+    io_set_output(STEPPER4_MSTEP0, (STEPPER4_MSTEP & 0x01));
+#endif
+#ifdef STEPPER4_MSTEP1
+    io_set_output(STEPPER4_MSTEP1, (STEPPER4_MSTEP & 0x02));
+#endif
+#ifdef STEPPER5_MSTEP0
+    io_set_output(STEPPER5_MSTEP0, (STEPPER5_MSTEP & 0x01));
+#endif
+#ifdef STEPPER5_MSTEP1
+    io_set_output(STEPPER5_MSTEP1, (STEPPER5_MSTEP & 0x02));
+#endif
+#ifdef STEPPER6_MSTEP0
+    io_set_output(STEPPER6_MSTEP0, (STEPPER6_MSTEP & 0x01));
+#endif
+#ifdef STEPPER6_MSTEP1
+    io_set_output(STEPPER6_MSTEP1, (STEPPER6_MSTEP & 0x02));
+#endif
+#ifdef STEPPER7_MSTEP0
+    io_set_output(STEPPER7_MSTEP0, (STEPPER7_MSTEP & 0x01));
+#endif
+#ifdef STEPPER7_MSTEP1
+    io_set_output(STEPPER7_MSTEP1, (STEPPER7_MSTEP & 0x02));
+#endif
+
+#ifdef ENABLE_PARSER_MODULES
+    ADD_LISTENER(gcode_parse_delegate, m351_parse, gcode_parse_event);
+    ADD_LISTENER(gcode_exec_delegate, m351_exec, gcode_exec_event);
+#endif
+}
 #endif
